requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.11
djangorestframework==3.9.0
idna==2.8
mccabe==0.6.1
nltk==3.4
pycodestyle==2.4.0
pyflakes==2.0.0
pytz==2018.7
requests==2.21.0
singledispatch==3.4.0.3
six==1.12.0
textblob==0.15.2
urllib3==1.24.1
